Create Game Min DTO.

+++ listgame/src/main/java/com/rodi/listgame/entities/Game.java
@@ -16,7 +16,7 @@ public class Game implements Serializable{
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String title;
 
@@ -37,7 +37,7 @@ public class Game implements Serializable{
     public Game() {
     }
 
-    public Game(Integer id, String title, String year, String genre, String platforms, Double score, String imgUrl,
+    public Game(Long id, String title, String year, String genre, String platforms, Double score, String imgUrl,
             String shortDescription, String longDescription) {
         this.id = id;
         this.title = title;
@@ -50,11 +50,11 @@ public class Game implements Serializable{
         this.longDescription = longDescription;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
